Moved implementation into implementation file
