+++ src/plugins/lua/MQ2Lua.vcxproj.filters
@@ -48,6 +48,9 @@
     <ClCompile Include="LuaCoroutine.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="LuaPlugin.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="LuaThread.h">
